Move cert migration to GotifyApplication and synchronize calls
