+++ app/src/main/kotlin/com/github/gotify/GotifyApplication.kt
@@ -4,9 +4,13 @@ import android.app.Application
 import android.app.NotificationManager
 import android.os.Build
 import androidx.preference.PreferenceManager
+import com.github.gotify.api.CertUtils
 import com.github.gotify.log.LoggerHelper
 import com.github.gotify.log.UncaughtExceptionHandler
 import com.github.gotify.settings.ThemeHelper
+import java.io.File
+import java.io.FileOutputStream
+import java.io.IOException
 import org.tinylog.kotlin.Logger
 
 class GotifyApplication : Application() {
@@ -26,6 +30,24 @@ class GotifyApplication : Application() {
             )
         }
 
+        val settings = Settings(this)
+        if (settings.legacyCert != null) {
+            Logger.info("Migrating legacy CA cert to new location")
+            var legacyCert: String? = null
+            try {
+                legacyCert = settings.legacyCert
+                settings.legacyCert = null
+                val caCertFile = File(settings.filesDir, CertUtils.CA_CERT_NAME)
+                FileOutputStream(caCertFile).use {
+                    it.write(legacyCert?.encodeToByteArray())
+                }
+                settings.caCertPath = caCertFile.absolutePath
+                Logger.info("Migration of legacy CA cert succeeded")
+            } catch (e: IOException) {
+                Logger.error(e, "Migration of legacy CA cert failed")
+                if (legacyCert != null) settings.legacyCert = legacyCert
+            }
+        }
         super.onCreate()
     }
 }
