+++ app/src/main/kotlin/com/github/gotify/api/ClientFactory.kt
@@ -7,10 +7,6 @@ import com.github.gotify.client.api.UserApi
 import com.github.gotify.client.api.VersionApi
 import com.github.gotify.client.auth.ApiKeyAuth
 import com.github.gotify.client.auth.HttpBasicAuth
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import org.tinylog.kotlin.Logger
 
 internal object ClientFactory {
     private fun unauthorized(
@@ -60,23 +56,6 @@ internal object ClientFactory {
         baseUrl: String = settings.url
     ): ApiClient {
         val client = ApiClient(authentications)
-        if (settings.legacyCert != null) {
-            Logger.info("Migrating legacy CA cert to new location")
-            var legacyCert: String? = null
-            try {
-                legacyCert = settings.legacyCert
-                settings.legacyCert = null
-                val caCertFile = File(settings.filesDir, CertUtils.CA_CERT_NAME)
-                FileOutputStream(caCertFile).use {
-                    it.write(legacyCert?.encodeToByteArray())
-                }
-                settings.caCertPath = caCertFile.absolutePath
-                Logger.info("Migration of legacy CA cert succeeded")
-            } catch (e: IOException) {
-                Logger.error(e, "Migration of legacy CA cert failed")
-                if (legacyCert != null) settings.legacyCert = legacyCert
-            }
-        }
         CertUtils.applySslSettings(client.okBuilder, sslSettings)
         client.adapterBuilder.baseUrl("$baseUrl/")
         return client
